backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
matplotlib
openai
google-generativeai
duckdb